requirements.txt
bert-extractive-summarizer==0.4.2
torch==1.5.0